Finished before flight test...I think
There are some hacks in this, due to running out of 0-254 indexed characters in the font, needed for the char array.

+++ src/main/io/osd_dji_hd.c
@@ -749,7 +749,7 @@ static void osdDJIFormatDistanceStr(char *buff, int32_t dist)
             else {
                 // Show miles when dist >= 0.5mi
                 tfp_sprintf(buff, "%d.%02d%s", (int)(centifeet / (100 * FEET_PER_NAUTICALMILE)),
-                (abs(centifeet) % (int)(100 * FEET_PER_NAUTICALMILE)) / FEET_PER_NAUTICALMILE, "NM");
+                (int)((abs(centifeet) % (int)(100 * FEET_PER_NAUTICALMILE)) / FEET_PER_NAUTICALMILE), "NM");
             }
             break;
         case OSD_UNIT_METRIC_MPH:
